Bug#35228656 Clean up usage of 'xfrm' flag in multiple 'read attribute' / 'read key' methods
Starting from Mysql-8.0, the 'xfrm' of a key is only used when we need to
generate a hash value from the key, not for a 'compare' any more.

However, there are still many methods used only for reading keys for compare,
or even general methods reading any kind of attributes, which takes as
'bool xfrm' argument.

This opens up for incorrect usage of these methods and should be avoided.

This patch removes that 'xfrm' argument from methods where it should
always be 'false'. It also cleans up the naming of other methods
used only to read keys for hash generation by adding a '*Hash*' to
their name.

Furthermore, the method Dbtup::readKeyAttributes() is introduced as
a replacement for calling ::readAttribute(). The latter method added
an AttributeHeader to the values being read, which is not needed
when used from Dbtup::tuxReadPk(). Thus it had to remove that
AttriButeHeader, requiring an extra copy of the PK.

Change-Id: I4daf3ee959c4b6a9f6fbbaebf5cc5bb0ffd6bbfb