Bug#35208539:Server abort from test_if_ref sql/sql_optimizer.cc:8917
If an error is returned while initializing the planner arrays
used in join order optimization, optimizer does not reset the
JOIN_TAB structure in TABLE object. This results in problems
for the next query that uses the same setting in TABLE structure
leading to server exit.
Fix is to reset "join_tab" in TABLE::reset.

Change-Id: I2a18a78d14082b3776ecc91943c097c4e1bce0df