Bug#35303407 MySQL Command Line Client acts on content between dollar quotes
Let the MySQL Command Line Client handle dollar quotes the same way
as ordinary quotes.

This is necessary in order to avoid that the client modify the body
of MLE routines. Otherwise, the client may act on, e.g., single or
double quotes, backslash characters, or character sequences that may
be interpreted as an SQL comment (# or --).

If a dollar quote is detected (see WL#15254 for details), the client
will enter string mode, and the prompt will be changed to '$>'.
When the matching dollar quote is found, one returns to normal mode.

An additional benefit of this change is that it will not be necessary
to change the client-side statement delimiter for MLE routines.

Change-Id: I1fc6878e32d2320f022f2e3f016a3fcfa5eef65b