Bug#35228076 MySQLSession::set_ssl_options sets object state that is uninitialized otherwise.
MySQLSession::set_ssl_options() sets the ssl-related session options on the
session object but also caches them. If set_ssl_options() was never called, the
cache contains uninitialized (partially random) values which may lead to errors
in the using code.

This patch changes the implementation so that the SSL options are never cached.
If needed they are queried from the underlying MYSQL object to reflect the
actual values.

Change-Id: I4b75111fbefab5488e2a2f1e24364ff549651f03