Bug#354288890: Histogram commands invalidate TABLE_SHARE
Post-push fix.

The use of an unreferenced named parameter in the catch-clause of a
try-catch block resulted in an "unreferenced local variable" warning
when building on windows.

This was fixed by changing the catch-clause into the form that declares
an unnamed parameter in two cases where the named parameter was
unreferenced, so instead of writing "catch (const std::exception &e)" we
write "catch (const std::exception &)"

Change-Id: I53f2dc2757ab36a9e0a5fea5238676cf81ebd1e8