Bug#35237765 MTR: Decouple port allocation for secondary engine processes from venv setup
Reservation of ports for secondary engine processes (i.e.,server,
driver) is being done in the routine used to create and setup a
virtual environment. Since port allocation does not depend on the
environment, the patch moves it out to a separate routine and adds
a way to exercise the situation when the environment variable
MTR_SKIP_VIRTUAL_ENV_SETUP is set.

Change-Id: Ib96c5aabb2a36199b0f063b63882d035c9a22639