Bug#35239607 Lost connection when max_connections reached
When a client connects through Router and max_connection
is reached at the server (and a SUPER connection is already
connected to the server), the client gets

  Lost connection to MySQL server during query

or similar.

Background
----------

The client received a Error-Message which was in the 3.21 format
(without SQL-state) even though it expected a Error-message in 4.1
format (with SQL-state). The client discarded the wrong message and
dropped the connection.

Additionally, the error-message was sent twice.

Change
------

- don't forward the Error packet AS IS, but call the 'on-error' handler
  to let it handle it by either recoding the error-msg according to the
  client's capabilities, or retry
- added tests for max-connection errors

Change-Id: I08c278ac342be008d651fad42a8a9ac85082a40f