Bug#35041573: Server diagnostics area error:
Problem:
When we execute FLUSH BINARY LOG or SET GLOBAL CHECKSUM=CRC32 after setting the
session transaction mode to READ ONLY, there is a crash.

Analysis:
When we execute the above two statements it tries to rotate the binary log, but
before rotating binary log we need to update the mysql.gtid_executed table.
Now because the session transaction mode was set to READ ONLY the open_table
command fails. But rather than returning the error we clear it and set the
dignostic area to the status DA_OK in the save method inside rpl_gtid_persist.cc file

int Gtid_table_persistor::save(const Gtid_set *gtid_set, bool compress) {
....
   /*
     Gtid table is not ready to be used, so failed to
     open it. Ignore the error.
   */
   thd->clear_error();
   if (!thd->get_stmt_da()->is_set())
     thd->get_stmt_da()->set_ok_status(0, 0, nullptr);
....
}

and return from this method assuming everything went fine. Now at the end of
execution(in mysql_execute_command) the code tried to set the status to DA_OK
in the my_ok method after the successfull return but crashed inside  my_ok(It was expecting the status
to be DA_EMPTY) method.

For checksum:

The method global_update for checksum didn't take into account the return value of rotate,
so upon returning to the mysql_execute_command, it was treated as there was no
error and again upon calling my_ok() the crash happened.

Fix:

Allow the rotate command by ignoring the READ ONLY transation mode as it is done
in the case of read-only and super-read-only server mode.

Change-Id: Ia61212325c5617510d2a67342cd46af9b40ca51f