Bug#35185670 mysqld crashes on expensive query
Addendum patch fixing build break of SectionIterators-t
on Sparc, 7.5 & 7.6 only.

Studio 12.6 Sun 5.14 is used to build and link the
unit test binaries for SectionIterators-t. The linker
seems unable to remove references to methods defined in
included header files, but not really used by the source
files SectionIterators.cpp and NdbApiSignal.cpp.

Thus we had build breaks due to 'undefined symbols' from the
linker.

Patch include the libraries ndbgeneral ndbportlib when linking
SectionIterators-t.

Intended for 7.5 and 7.6 only as 8.0++ use gcc
which is not affected by this problem.

Change-Id: I099ff525856cfa7ad9d3246bdc61c36af75aaaa8