Bug#35149996 ndb_binlog_log_apply_status.test does not test what it is intended to
Test case insert two new rows into the ndb_apply_status table and expect
these inserts to be written to the binlog.

As updates to the ndb_apply_status table is excemptet for what is
counted as an update within an epoch, the test case also updates
the t1 table in order to create a non empty epoch.

However, the t1 update is to the exact same value as the existing one,
thus the update is ignore and the epoch is still empty
 -> No ndb_apply_status updates written to th ebinlog.

Patch change the t1 update operation, such that a real update takes
place

Change-Id: I8a5cdfc5ddf7e4638f7959020b981704cf1db039