Bug#34897517: Can't set report_host to NULL
Setting variable values on the command line doesn't support DB NULL
values.
Hence it should not be allowed to persist a DB NULL as a values for
one of the variables that takes its persisted values from the command
line at startup.

Added a check and an error message at SET PERSIST[_ONLY] time.
Added a list of exempted variables.

Change-Id: Ic95cd946314cd31fe2f0217d28a20b45ea674e3e