Bug#35254738 mtr_t::is_dirty() and mtr_t::Impl::m_made_dirty were never used for anything
Since its introduction in 2013, the mtr_t::is_dirty() was not really
used for anything, so let's remove this.
The m_made_dirty was maintained by checking if the mtr has modified a
page and if the page was dirtied earlier, in the hope that we could
avoid the whole step of updating flush lists at mtr commit.
Similar logic is already performed on page-granularity: we decide for
each page if it was dirtied by this mtr before adding this particular
page to a flush list.

Change-Id: Ie7763e60b13a288f94b4153fe4f6f899b693789f