Bug#35318797 decode StmtExecute after StmtParamAppendData fails
Sending StmtParamAppendData (aka COM_STMT_SEND_LONG_DATA) before a
StmtExecute currently fails (which is only visibible if the tracer is
enabled.)

StmtParamAppendData allows to send data to a prepared-statement
parameter before the statement is executed. Parameters of a statement
which already filled with StmtParamAppendData do not send data in
StmtExecute.

This requires tracking the StmtParamAppendData when StmtExecute is
called.

Change
------

- track which parameters of a statement already got data via
  StmtParamAppendData

Change-Id: I97ad0bc53da124b13fe00eef725852a0d39cb300