Bug#35259868 classic_protocol_server_greeting_error fails
routertest_integration_routing_direct and
routertest_integration_routing_sharing_constrained_pools fail
after

  Bug#35239607 Lost connection when max_connections reached

was pushed.

Background
----------

The test reduces the max-connections of the server and wants to reset it
after the test finished.

Given to timing the last connection of the test may still a "active"
when the cleanup code tries to open a new connection and reset the
max-connection value.

Change
------

- retry to connection in case error 1040 "Too many connections" for
  1 second

Change-Id: Ic84338aadc8734d1a75c960a2f3598a244859d39