Bug#35092279 Use of push_warning_printf with severity level SL_ERROR
push_warning / push_warning_printf do not accept ERROR severity in
messages since these are usually logged through other means and are
not presented to the user through the WARNINGS stack. If called thus
then it reduces the severity to WARNING before pushing and logging.

On DEBUG builds, however, it crashes with a failed assert.

This commit changes ndbcluster handler calls to push_warning_printf to
use severity WARNING and, additionally, leverages the var args of
string formats present in the function already.

Change-Id: I2b05523bee2755c5aea500ecaa1fad5ebc016128