Bug#35101169 Assertion 'nullptr != dynamic_cast<Target>(arg)' failed Bug#35150085 Wrong item name in result set after transform
A case of transforming subqueries to a JOIN with derived tables, with
the containing query being grouped. We create an extra derived table
in which to do the grouping. This process moves initial select list
items from the containing query into the extra derived table,
replacing all original select list items (with the exception of
subqueries which get their own derived tables) with columns from the
extra derived table.

This logic didn't handle the DEFAULT function correctly:
Item_default_value isn't modelled interally as a function, but is
rather a subclass of Item_field. We hit an assert when trying to cast
an Item_default_item as an Item_view_reference: the item in question's
type was not ITEM_FIELD, or ITEM_VIEW_REF as expected but DEFAULT,
which wasn't handled.

The patch adds support for DEFAULT(expression) in queries undergoing
the mentioned transform.

This patch also solves an issue with wrong item_name in the metadata
in a corner case: two occurrences of the same column in the select
list got the same Item::item_name as a result of the transform.

Change-Id: I8ced367145b40bd160873bbf45d2bcd468dd28ce