Bug#35259704 	Windows build with -DWIN_DEBUG_NO_INLINE=1 fails on "library limit of 65535..."
The mysqld.def file is too long, it can't have > 65535 entries.
Last time the same problem was improved it was in Bug#33357465, where we disabled std::forward template method instances.
However, we did not check if it is in the `std::` namespace, and were cutting out possibly all "forward" methods.
Also, we really don't need any std:: methods exported, as they are only part of the standard C++ library.
To remove all the std:: methods, we will check unmangled part of the `symbol_line` to look for " __cdecl std::", where the `__cdecl` is function call type modifier which is followed by the name with namespaces. I hope this will not be present in methods which are not in `std::` namespace, but have a template argument to such method.

This brings down the number of symbols exported to ~38300, leaving a lot of room for expansion.

Change-Id: I9ca13beba02adbfe4daa7d452ad0be803d190440