introduce SCHED TRANSCATION command
This commit changes the parser to support a new way to start a
transaction---the SCHED TRANSACTION command. The syntax is

```
SCHED TRANSACTION [transaction_type] [transaction_args]
```

This command is identical to START TRANSACTION (or BEGIN), except it
first buffers the transaction by calling a storage engine handler
function. The actual buffering (scheduling) is not implemented yet.

The purpose of the transaction type and the transaction args is to
enable us to map a transaction to a scheduling cluster. The type is an
unsigned integer, and the arguments are a list of literals.