BUG#35244143 - Unsetting items_per_page on MRS Schema and DB Object crashes the router
Decription
==========
When a REST object doesn't have `items_on_page` parameter configured, on some
platforms it leads to a crash. "Doesn't have configured" means that in DB_OBJECT
and its DB_SCHEMA the parameter is set to NULL.

Fix
===
The implementation of `std::stoul` differes on some platforms, causing a crash
when converting NULL value or an exception is thrown.
The fix, add a check for the NULL value and allows the caller to define
a default value for the paramter in case when the DB did return a NULL.
The default value for `items_on_page` was set to 25 (like a default in the
schema).

RB: