Bug#35208605: Mysqldump table with virtual column wrong syntax
The fact that some columns aren't printed in the INSERT statements
generated by mysqldump wasn't taken into account when deciding on
printing the commas for the INSERT VALUES list.
Fixed by having a special flag if a column is already printed or not
instead of deciding based on column index.

Test cases added to cover all kinds of virtual columns.

Change-Id: I55484c3a795a4ad18b24c0691b2def9a04306e7a