Bug#35277644: Fill MEM_ROOT allocated memory with garbage in debug mode
Fill MEM_ROOT allocated memory with garbage, using the TRASH macro, on
allocation and deallocation, so that it is easier to detect bugs that
involve reading uninitialized memory.

One unit test had to be updated because its cleanup code could end up
destroying the same Fake_TABLE object multiple times, and overwriting
it with garbage on the first destruction made the second destruction
fail. The unit test was changed to destroy each object exactly once.

Change-Id: Ifb1f5b1ba429ac83c2394da7ee3092deb57a560b