BUG#34815537: kill connection when executing group_replication_switch_to_single_primary_mode
When a group action is sent to the group and the connection is killed on
coordinator, members will be on different states. Members that received
coordination action are waiting for member that executed action, and
member that started execution don't have nothing to process.

This cause a problem on coordination on the group.

To prevent this group actions sent to the group will wait until are
finished.

Change-Id: Iec331702e2c1a547afe18c866148cec820dec690