Bug#35283401 Clean up cmake code for MSVC code analysis
This is a followup to the patches for
Bug#34828882 Enable MSVC Code Analysis for C/C++

Do the check for MSVC warnings similarly to our "maintainer" mode,
i.e. *after* all 3rd party configurations are done.

Do not disable the MSVC CPPCHECK in googletest.cmake as this
will effectively disable it globally. Add an explicit suppresion
for /wd26433 for googetest libraries instead.

Rename the MSVC_CPPCHECK macro to DO_MSVC_CPPCHECK, as it was a bit
confusing to have an option and a macro with the same name.

Add missing call to MSVC_CPPCHECK_DISABLE in keyring_aws/CMakeLists.txt

Change-Id: Id41395dd7fdcd30f4098cbb757e5b7215d5db337