Bug#34826692 - Index merge should favour union over sort_union
Issue:
Index merge should favour "ror-union" over "sort-union" plans, if
they have similar costs, since "sort-union" requires sorting the
rows by rowid's whereas the "ror-union" does not.

Analysis:
If a WHERE clause has an "OR" condition, for each part of an "OR"
condition, range optimizer gets the best range scan possible and
uses all these range scans to build an index merge scan i.e a
"sort-union" scan. However, if it finds that all the best range
scans are also "ror-scans" i.e rows retrieved from these scans come
rowid ordered, then range optimizer always proposes a "ror-union"
scan because it is always cheaper than a "sort-union" scan.
Problem arises, when the best range scan for any one part of an
"OR" condition is not a "ror-scan". In this case, range optimizer
always picks a "sort-union". However, in some cases, it could be
advantageous to pick a "ror-scan" even though it might mot be the
best range scan to handle one part of an OR condition because it
would avoid sorting by rowid's.

Fix:
When finding the best range scan, range optimizer also detects if
there was a possible "ror-scan". We use this information to see if
each part of the OR condition has at least one possible "ror-scan".
If so, we rerun the range optimizer to get the best "ror-scan" to
handle each part of the OR condition and make a "ror-union" path.
We compare this cost with the "sort-union" cost to propose the
final plan.

Change-Id: Ibe52a90cae03878c453b99f7292a8989c277edfd