Bug#35289390 purge_sys->thds.find(thd) is causing congestion
Instead of storing thd's of purge threads in an unordered_set protected
by the congested purge_sys->latch, use a thread_local bool to indicate
if a thread is related to purge or not.

This lets innobase_trx_allocate() avoid S-latching the purge_sys->latch
which might be X-latched by actual purge coordinator thread, while he is
busy waiting for trx_sys->mutex to copy the oldest readview, which
apparently was the scenario leading to "long semaphore wait".

Change-Id: I2397f0fed0817badbaec74c5561141cbb1627efc