BUG#35292716 Unnecessary alignment check for application pointer in NdbEventOperation
Problem:

While configuring which attributes to receive while setting up a
NdbEventOperation it's possible to pass a pointer to buffer provided by
the application. When data is later received it should be available in
that specified location.

The received data is currently properly placed in the provided buffer
location, but due to check of pointer alignment as well as check of
attribute bytes size the NdbApi also allocates internal buffers.

The problem primarily manifest itself in applications that subscribe to
data changes from NDB using the NdbEventOperation::getValue() and
NdbEventOperation::getPreValue() functions with buffer provided by
application. One such application is the ndb_binlog thread which
subscribes to data changes to be delivered directly to a MySQL format
"record" buffer. In that case each field has its location in the buffer
determined by the table definition and in most cases it will not be
aligned or have an even field length, thus the NdbApi internal buffers are
allocated needlessly in many cases.

Analysis:

The allocated internal buffers are not used, this is wasteful. Presumably
the alignment and byte size check is a legacy requirement from when
received data was copied word by word, nowadays memcpy() is used to copy
data to correct place and thus there is no need for aligned pointer or
attribute of specific size.

The problem can be detected by comparing the pointer returned from
NdbRecAttr::aRef() with the pointer passed to getValue() or
getPreValue() functions, when no internal buffers are allocated they
should be identical.

Fix:

Remove the check for word aligned pointer and attribute byte size being
an even multiple of word size, thus avoiding that internal buffers are
allocated.
Remove unnecessary pointer comparisons when receiving data since the
requirement is that data should always be copied (using memcpy()) to
location provided by pointer "theRef", thus removing some checks from
presumably hot code path.

Change-Id: I0e852f5e7c4a1562109d0e2df38e36359d9e5804