Bug#34826692 - Index merge should favour union over sort_union
POST-PUSH FIX:
main.index_merge_innodb test was failing.

ISSUE:
When we rerun the range optimizer to get best "ror-scan" to
make a "ror-union" path, range table read plan could not be
found as not a single key got selected since we are considering
only "ror-scans". Later while finding a "ror-union" path this
read plan was getting accessed which was NULL.

FIX:
This patch will fix the issue by checking if read plans
are found before checking for "ror-union" path.

Change-Id: Ie5ccc13c591b7b0da8e57fe394a7a5494544d298