Bug#34826692 - Index merge should favour union over sort_union
POST-PUSH FIX:
Some tests were failing on daily trunk.

ISSUE:
When we rerun the range optimizer to get best "ror-scan" to
make a "ror-union" path. Here if optimizer trace is enabled it
tries to add an object member to the current trace context
which was not present.

FIX:
Call the optimizer trace object before finding "ror" read plans.

Change-Id: I072d72f96e03611d71ee1a43c1589b848eed5b43