Bug#35302794 - Segfault in get_best_disjunct_quick on JOIN and                range predicates
This is a regression due to patch for Bug#34826692 where rerun
the range optimizer to get best "ror-scan" to make a "ror-union"
path.

After calculating the cost for "ror-union" path, we then
compared this cost with the "sort-union" cost to propose the
final plan. Here it tried to access the cost of "sort-union"
path which was NULL.

The fix is to check if "sort-union" path is found before comparing
the costs.

Change-Id: Ide8382413900cb4069a76a67a0a634c1ab13106c