Bug#34836463 Data node should print stack trace only for software error cases Bug#34629622 Stack trace shown twice when data node process receives unix signal
When a data node is shut down with an error it is produced a stack
trace on stdout. This can result in a stack trace being generated
for normal 'controlled exit with error' cases, and this gives
the impression of a crash, requiring a bug report, follow up etc.

To prevent these situations in this patch the stack trace generating
code is refined to produce stack traces only in unexpected situations.
In the code the 'unexpected situations' are called as software
error.
Currently only the errors with ndbd_exit_classification XIE
(internal_error) are considered as 'software error'.

This patch also prevents the stack trace to be shown twice when data
node process receives unix signal. (Bug#34629622)
In ErrorReporter::handleError(), for the NDBD_EXIT_OS_SIGNAL_RECEIVED
error code, although it is a software error (ndbd_exit_classification
XIE), stack trace is not printed because it is already done in the
Unix signal handler case.

Change-Id: Ia7c0afa2c1093ae74b12a09824e1c3de88125d3f