Bug#34361437 SELECT DISTINCT handles 0 vs. -0 differently in OO and HG
For GROUP BY or for DISTINCT, the temporary table uses hash table for
deduplication. The hash value is calculated from binary data values.
So for approximate types such as float or double, 0.0 and -0.0 hash to
two different values because their binary representation is different.
As a result, 0.0 and -0.0 were treated as two distinct values.  Fixed
this by converting -0.0 to 0.0 before calculating the hash.

While the above issue arises in Cell_calculator::hash(), a similar
issue arises in Aggregator_distinct::setup() during calls such as
COUNT(DISTINCT) or AVG(DISTINCT). Fix this by *not* using a raw binary
comparison function for approximate types when initiating the 'Unique'
object.

Change-Id: I310f1a7b322d0576b70b683ce168dcae901f711f