Bug#35327995 mysql build error with gcc12
Post-fix for Bug#34361437 SELECT DISTINCT handles 0 vs. -0 differently.

g++-12 compiler with -O2 gives a false alarm with an
array-out-of-bound error.  While compiling the unit test for
cell_calculator-t.cc, g++-12 is able to identify that the cell data is
2 bytes and that there is a call to float8get() which accesses 8
bytes. But it is not able to identify that float8get() will be called
only if the cell data is 8 bytes, incorrectly leading to array bound
errors.

In cell_calculator-t.cc, allocate the cell buffer of size 8 bytes to
make the compiler happy.

Change-Id: Id33e0aaacb2af4d713f1e5e4c5cd4bce13e9c976