Bug#35178672: Assertion `dptr <= buffer->ptr() + buffer->alloced_length()' failed.
Assert failures were sometimes seen with the hypergraph optimizer in
queries that performed covering index scans referencing virtual
generated columns.

The problem was that the row buffers used for sorting, hash joins and
streaming aggregation would attempt to store every column in
table->read_set. When accessing virtual generated columns, the read
set contains the base columns on which the virtual column depends,
even if the base columns are not explicitly referenced by the query.
When using covering index scans, those base columns are not
necessarily available in the covering index, and attempting to read
them into the buffer does not make sense and could fail.

In the old optimizer, this is accounted for by removing the extra
columns from the read set if a covering scan is chosen.
DynamicRangeIterator and AlternativeIterator have additional handling,
since they can switch between covering and non-covering scans during
execution, and swap between full read set and reduced read set
depending on which kind of scan is actually chosen. In addition to
being complicated, this is unnecessary. Even if a covering index is
not used, the base columns will never be read from the buffer if they
are not referenced by the query, so storing only the columns in the
reduced read set in the buffer is fine regardless of access method.
And storing fewer columns in the buffers should be more efficient,
also in the non-covering case.

The fix is therefore to remove the current code that modifies the read
set for covering index scans, and instead store an additional read set
in the TABLE object, without the non-referenced base columns of
virtual generated columns. The buffers in pack_rows::TableCollection,
used by streaming aggregation and hash join, and the buffers in
Filesort are always filled with the columns referenced in the
alternative read set. This avoids the attempts to read columns not
available in a covering index scan, regardless of which optimizer is
used.

Change-Id: Ic3d906c22b23e63c12eadfc5f39e8be505d0a446