BUG#33650776 Failure of XA COMMIT of prepared txn, can result in txn rollback
Issue description
-----------------

In the context of WL#14700and while inspecting why
`mysql-test/suite/binlog_nogtid/t/binlog_gtid_next_xa.test` would fail with
the `xa_detach_at_prepare = ON`, found that there is disparity in error
treatment between attached and detached XA transactions.

Analysis
--------

Detached and attached transactions are treated differently when the
transaction is rolled back by the server after prepare, when some issue is
encountered during the `XA COMMIT` execution.

For attached transactions, if an issue is found after prepare and the
transaction is rolled back, the `XA ROLLBACK` event is not written to the
binary log and the transaction is kept in `ROLLBACK_ONLY` state, which
means that will appear in `XA RECOVER` but can only be rolled back, not
committed. If a restart/crash happens in the meanwhile, the transaction
will no longer appear in `XA RECOVER`.

For detached transactions, if an issue is found after prepare and the
transaction is rolled back, the `XA ROLLBACK` event is written to the
binary log and the transaction is finished, it won't appear in `XA
RECOVER`.

Proposed solution
-----------------

1. Make the behavior the same for both attached and detached transactions.
2. Do not rollback the transaction if an issue occurs after prepare, leave
  it prepare state and return XA_RETRY (ER_XA_RETRY in MySQL).

Change-Id: I7dadba2f5f6baaa13a020a66cd35c8c5c8ffd791