WL#15628: Adding support for new string functions in HeatWave
-> BIN()
-> ELT()
-> EXPORT_SET()
-> FIELD()
-> MAKE_SET()

Bug#35604335: Mysqld crash - Assertion `!hasRoot_' failed
Bug#35625941: Rpdserver crash - sig11 at bv_bittest|if/bitvector.h
Bug#35645004: ELT ASAN bug access out of bounds

** Bug#35604335 **
In some cases where data where not supported for BIN() during the
primitive setup, and thus the operator was not created correctly we
returned FALSE instead of TRUE causing issues in the formed QKRN

** Bug#35625941 **
In the FIELD() primitive we encountered issues in the nullability of
data that can cause the has_nulls constexpr to be TRUE where
col->has_nulls_rpdmqx_col is FALSE and thus the
nullbitvec_startidx_rpdmqx_col is NULL, so when accessed during
is_null() we crush. This happens because we wrongly overwrite the
nullability of intermediate operators applied to the input columns
of FIELD() during QKRN creation. Removed the erroneous overwrites.

** Bug#35645004 **
One of the input arguments of the ELT primitive is the previous result
as it executes multiple rounds accumulating results at each round.
In the first round, the previous result is invalid and thus we should
not increase its iterator.

Change-Id: I2baa406f7c8dc86fed055583fc9ff8d2b22e761a