Bug#35338776: Failure in find_item_in_list with unknown column in field list
The problem is that when a column reference given by table name and
column name is looked up in the function find_item_in_list(), we
ignored that the searched item may not have a table name, as it is
not yet resolved. Now, we should really only search for strict alias
names in find_item_in_list(), however removing this support would be
a much larger change, so we restrict the fix to an explicit check for
a null table name in the searched item.

Change-Id: I216f478efe72545e30c61e56919419db8b81f60d