Bug#35341198 StatsUpdatesFrequencyTest fails randomly on pb2
The test assumed that the Router is running for 2 seconds.
The math in the test body was off, it was waiting for 20 transactions
each about every 50ms which is 1s+ not 2 seconds.

This patch replaces wating based on the Router metadata transactions
with direct 2 seconds wait, which makes more sense as the tested
functionality also is based on the steady_clock duration (with ~50ms
resolution from metadata.ttl)

Change-Id: I76f92845bd0af9253c241ac672f1ff99a3c4a750