Bug#35340987 [backport] Prepared SELECT returns different results on consecutive executions.
This patch also solves:
Bug#35380604 issue with window functions in stored procedures
Bug#35846873 Segmentation fault in /mysql-8.0.34/sql/item.cc:2811
Bug#35846585 Arithmetic exception in /mysql-8.0.34/sql/item_sum.cc:5032

A regression from

  Bug#35060385 Item::update_used_tables should also update the PROP_SUBQUERY

which fully re-calculates m_accum_properties for an Item_func,
starting with 0.

In the repro, we have a prepared statement which gives the wrong
result in the second execute.  We have an equality expression in the
select list, one argument of which is a window function. During
optimize, we find ourselves in this situation:

    Item_ref::update_used_tables   (Item_aggregate_ref to w.f.)
    Item_func::update_used_tables  (equality function)
    ReplaceMaterializedTable
    CreateFrameBufferTable
    make_tmp_tables_info
    :

But make_tmp_tables_info changes the slice for the window function
with this call: set_ref_item_slice(widx). So Item_ref's m_ref_item
then points to a field in a window tmp table, not the window function,
so we lose the PROP_WINDOW_FUNCTION property for the aggregate
reference and hence also the equality function, which leads to error
in the second execute of the prepared statement: since optimize now
finds the equality function now contains no window function it is
evaluated too early: during a call of copy_funcs from
BufferingWindowIterator::Read instead of during a call of copy_funcs
from process_buffered_windowing_record.

The solution is always (re)compute update_used_tables for the function
fields in make_tmp_tables_info if windows are present, i.e. before we
start relying on the correct values of the PROP_WINDOW_FUNCTION
property for the (next) execution.

Change-Id: I43617988424ea3fc17c0cca5d5e5e40a86b46b8c