Bug#35311140 Build failure with Homebrew rapidjson installed & symlinked
On MacOS, If Homebrew rapidjson is installed, then build fails. Note that
even -DWITH_RAPIDJSON=bundled does not help, and the only workaround is brew
unlink rapidjson.

This is a regression from
    Bug#35006191 Convert rapidjson to a cmake interface library
which removed the global
    INCLUDE_DIRECTORIES(BEFORE SYSTEM ${RAPIDJSON_INCLUDE_DIR})

It turns out we need this on MacOS to avoid accidental #include of
rapidjson headers from Homebrew rather than the bundled code in
extra/rapidjson.

Change-Id: Ie656f064e21b031a2d4716427e4d2a4440fba7d7