Bug#35082046: AccessPath* {anonymous}::CostingReceiver::ProposeAccessPath Assertion `false' Bug#35096962: Assertion : `false' in (anonymous namespace)::CostingReceiver::ProposeAccessPath
The hypergraph optimizer could in some situations push down
non-deterministic predicates from the WHERE clause to join conditions.
We usually don't push down non-deterministic predicates, since they
should be evaluated at the latest possible point to avoid caching the
values (see the description of RAND_TABLE_BIT in sql_const.h).

Pushing down such predicates sometimes caused assert failures because
of inconsistent row estimates when the hypergraph was cyclic. The
inconsistencies happened because the selectivity of the
non-deterministic predicate would be counted early if a join with a
non-deterministic join condition was explored, whereas it would be
counted late if an alternative deterministic join was explored.

This patch makes PushDownCondition() stop pushing down
non-deterministic predicates to join conditions. It still allows
partial pushdown of the deterministic parts of a non-deterministic
predicate, but the full predicate itself needs to stay in the WHERE
clause.

Change-Id: Ia90dec537fb78b6f8e3794b147cb1f631e3a3535