Bug#34844509: Assertion `receiver.HasSecondaryEngineCostHook()' failed
When a query has so many interesting orderings to consider that it
doesn't produce a complete state machine,
LogicalOrderings::DoesFollowOrder() may incorrectly report that a path
sorted on the GROUP BY ordering does not follow the GROUP BY ordering,
and as a consequence no valid plan is found for the GROUP BY step.

Fixed by checking ordering indexes for equality first, and only
calling DoesFollowOrder() if the indexes are not equal.

Change-Id: Ie6642211a7705deefa891fa4b3b6c36477904c70