Bug#35337357 routertest_integration_routing_sharing_restart fails with ASAN
The routertest_integration_routing_sharing_restart test failed
in tests which take a connection from the pool and try to write a TLS
message to the server.

SSL_set_msg_callback() takes the current connection() as "arg". When the
server-connection is moved to the pool and attached to another
connection, the callback-arg was still pointing to the earlier
connection which already be destructed, leading to a use-after-free that
ASAN detected.

Change
------

- after taking the connection from the pool, set the callback-arg of
  the openssl's SSL_set_msg_callback() to current connection.

Change-Id: Ia76a598e29241b9db72d4d89898ff0a7542d707e