Bug#35323398: Assertion 'scale >= 0 && precision > 0 and               scale <= precision' failed
The problem is with the resolve_type_inner() function for class
Item_func_nullif. If the original data type is a temporal type,
the type is adjusted to a string type. However, the result type
is not adjusted accordingly, which may cause later inconsistencies.

Fixed by setting result type to STRING_RESULT.

Change-Id: I91b43741ced9c0253b9e337e7170d8c93ac9c8f4